requirements.txt
asgiref==3.11.1
asyncpg==0.31.0
black==26.3.1
click==8.3.3
dj-database-url==3.1.2
Django==6.0.4
djangorestframework==3.17.1
dotenv==0.9.9
mypy_extensions==1.1.0
packaging==26.2
pathspec==1.1.1
platformdirs==4.9.6
psycopg==3.3.3
psycopg-binary==3.3.3
python-dotenv==1.2.2
pytokens==0.4.1
sqlparse==0.5.5
